metal-music-archive-api/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
cffi==2.0.0
click==8.3.1
contextlib2==21.6.0
cryptography==46.0.3
Django==6.0.1
django-ninja==1.5.1
django-ninja-extra==0.30.8
django-ninja-jwt==5.4.3
h11==0.16.0
iniconfig==2.3.0
injector==0.24.0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
